+++ private/childProcessPromise.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const { ChildProcess } = require('child_process');
+
 /**
  * Promisifies a Node.js child process.
  * @kind function
@@ -8,7 +10,12 @@
  * @returns {Promise<{exitCode: number, signal: string}>} Resolves the exit code if the child exited on its own, or the signal by which the child process was terminated.
  * @ignore
  */
-module.exports = function childProcessPromise(childProcess) {
+module.exports = async function childProcessPromise(childProcess) {
+  if (!(childProcess instanceof ChildProcess))
+    throw new TypeError(
+      'First argument `childProcess` must be a `ChildProcess` instance.'
+    );
+
   return new Promise((resolve, reject) => {
     childProcess.once('error', reject);
     childProcess.once('close', (exitCode, signal) =>