+++ test/index.js
@@ -6,6 +6,7 @@ const tests = new TestDirector();
 
 require('./replaceStackTraces.test')(tests);
 require('./cli/coverage-node.test')(tests);
+require('./private/childProcessPromise.test')(tests);
 require('./private/reportCliError.test')(tests);
 require('./private/semver.test')(tests);
 require('./private/sourceRange.test')(tests);